benchmark/prosody_validation/pyproject.toml
[project]
name = "prosody-validation"
version = "0.1.0"
description = "Prosody validation benchmark for testing SLM understanding of speech prosody"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Core dependencies
    "openai>=1.0.0",
    "aiohttp>=3.9.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "tqdm>=4.66.0",
    "python-dotenv>=1.0.0",
    
    # ASR - Whisper
    "openai-whisper>=20231117",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    
    # Audio processing
    "soundfile>=0.12.0",
    "librosa>=0.10.0",
    
    # ML/Stats
    "scikit-learn>=1.3.0",
    
    # Configuration
    "pyyaml>=6.0",
    
    # Logging and monitoring
    "colorlog>=6.7.0",
    
    # Retry logic
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]
